requirements.txt
# ── Data & Compute ──────────────────────────────
pandas==2.2.2
numpy==1.26.4
pyarrow==16.0.0          # Parquet read/write
fastparquet==2024.2.0    # Alternative Parquet engine

# ── Market Data APIs ────────────────────────────
alpaca-py==0.26.0        # Alpaca REST + WebSocket
polygon-api-client==1.14.0  # Polygon.io OHLCV
finnhub-python==2.4.20   # News + sentiment

# ── Technical Indicators ─────────────────────────
pandas-ta==0.3.14b       # 130+ indicators (MA, RSI, Bollinger, ATR etc)

# ── LLM Layer ───────────────────────────────────
anthropic==0.26.1        # Claude API

# ── Scheduling ──────────────────────────────────
APScheduler==3.10.4

# ── AWS ─────────────────────────────────────────
boto3==1.34.84           # S3 sync + Secrets Manager

# ── Dashboard ───────────────────────────────────
streamlit==1.35.0
plotly==5.22.0

# ── LEAN ────────────────────────────────────────
lean==1.0.214            # LEAN CLI

# ── Testing ─────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0

# ── Utilities ───────────────────────────────────
python-dotenv==1.0.1     # Local dev only (prod uses Secrets Manager)
loguru==0.7.2            # Structured logging
pydantic==2.7.1          # Data validation & dataclasses
httpx==0.27.0            # Async HTTP client
tenacity==8.3.0          # Retry logic with exponential backoff
rich==13.7.1             # Beautiful terminal output
typer==0.12.3            # CLI commands
